% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombiROC_long.R
\name{CombiROC_long}
\alias{CombiROC_long}
\title{Reshape CombiROC data in long format.}
\usage{
CombiROC_long(data)
}
\arguments{
\item{data}{a data.frame returned by load_data().}
}
\value{
a data.frame in long format
}
\description{
A function that simply wraps dyplr::pivot_longer() to reshape data in long format.
}
\details{
This function returns the data in long format (with 'Markers' and 'Values' columns)
}
\examples{
# To reshape demo_data in long format

## demo_data_long <- CombiROC_long(data = demo_data)

}
