% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiroc_select.R
\name{combiroc_select}
\alias{combiroc_select}
\title{Select Best Markers Using PCA}
\usage{
combiroc_select(data, case_class, n_cpus = NULL, n = NULL, tol = 0.1)
}
\arguments{
\item{data}{A data.frame returned by load_data().}

\item{case_class}{A character that specifies which of the two classes in the dataset is the case class.}

\item{n_cpus}{an integer that specifies the number of cores to allocate.}

\item{n}{an integer specifying the number of top best markers to retain.}

\item{tol}{a value passed to prcomp(), indicating the magnitude below which principal components should be omitted. (Components are omitted if their standard deviations are less than or equal to tol times the standard deviation of the first component.) See stats::prcomp documentation.}
}
\value{
a list containing: "PCA" the result of stats::prcomp() and "best_markers" a named numeric vector with marker names as names their importance in distinguishing the case class.
}
\description{
This function retrieves the best n markers of the dataset by training a logistic regression model with all the markers. It then performs Principal Component Analysis (PCA) to identify the most significant features in relation to the specified case class. The function ranks the markers by their weighted importance in the PCA and selects the top n markers.
}
\examples{

demo_data # combiroc built-in demo data (proteomics data from Zingaretti et al. 2012 - PMC3518104)

# To select the best 3 single markers for class A with 1 CPU.

best_markers <- combiroc_select(data = demo_data, case_class = "A",n_cpus = 1 , n = 3)
}
