% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Combiroc built-in demo data (proteomics data from Zingaretti et al. 2012 - PMC3518104)}
\format{
A data frame with 170 rows and 7 variables:
\describe{
\item{Patient.ID}{the ID of samples}
\item{Class}{the class of the samples: A-Healthy, B-AIH}
\item{Marker1}{the signal intensity value of Marker1}
\item{Marker2}{the signal intensity value of Marker2}
\item{Marker3}{the signal intensity value of Marker3}
\item{Marker4}{the signal intensity value of Marker4}
\item{Marker5}{the signal intensity value of Marker5}
}
}
\usage{
demo_data
}
\description{
A dataset containing signal intensity values of a 5-marker signatures for Autoimmune Hepatitis (AIH).
Samples have been clinically diagnosed as “abnormal” (class A) or "normal" (class B).
}
\keyword{datasets}
