% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_seurat.R
\docType{data}
\name{demo_seurat}
\alias{demo_seurat}
\title{Combiroc built-in demo SeuratObject (https://cran.r-project.org/web/packages/SeuratObject/SeuratObject.pdf)}
\format{
A subset of the PBMC3K SeuratObject (6 features)
}
\usage{
demo_seurat
}
\description{
A subset of the pbmc3k.SeuratData::pbmc3k.final object containing only the following genes:
'RBP7','CYP1B1','CD14','FCN1','NKG7' and 'GNLY'
}
\keyword{datasets}
