% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_unclassified_data.R
\docType{data}
\name{demo_unclassified_data}
\alias{demo_unclassified_data}
\title{A demo dataset with unlabeled proteomics markers}
\format{
A data frame with 20 rows and 7 variables:
\describe{
\item{Patient.ID}{the ID of samples}
\item{Marker1}{the signal intensity value of Marker1}
\item{Marker2}{the signal intensity value of Marker2}
\item{Marker3}{the signal intensity value of Marker3}
\item{Marker4}{the signal intensity value of Marker4}
\item{Marker5}{the signal intensity value of Marker5}
}
}
\usage{
demo_unclassified_data
}
\description{
Combiroc built-in unclassified demo dataA dataset containing signal intensity values of a 5-marker signatures for Autoimmune Hepatitis (AIH). This dataset has been obtained by randomly picking 20 samples from demo_data without their classification.
}
\examples{
demo_unclassified_data # combiroc built-in unclassified demo data
}
\keyword{datasets}
