% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load CombiROC data.}
\usage{
load_data(data, sep = ";", na.strings = "")
}
\arguments{
\item{data}{the name of the file which the data are to be read from.}

\item{sep}{the field separator character.}

\item{na.strings}{a character vector of strings which are to be interpreted as NA values.}
}
\value{
a data frame (data.frame) containing a representation of the data in the file.
}
\description{
A customized read.table() function that checks the conformity of the dataset format, and only if all checks are passed, loads it.
The dataset to be analysed should be in text format, which can be comma, tab or semicolon separated:
\itemize{
\item The 1st column must contain patient/sample IDs as characters.
\item The 2nd column must contain the class to which each sample belongs.
\item The classes must be exactly 2 and they must be written in character format.
\item From the 3rd column on, the dataset must contain numerical values that represent the signal corresponding to the markers abundance in each sample (marker-related columns).
\item Marker-related columns can be called 'Marker1, Marker2, Marker3, ...' or can be called directly with the gene/protein name, but "-" is not allowed in the column name.
Only if all the checks are passed, it reorders alphabetically the marker-related columns depending on marker names (necessary for a proper computation of combinations), and it forces "Class" as 2nd column name.
}
}
