% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_overview.R
\name{markers_overview}
\alias{markers_overview}
\title{Show an overview of the expression of each marker.}
\usage{
markers_overview(data_long, ylim = NULL)
}
\arguments{
\item{data_long}{a data.frame in long format returned by CombiROC_long()}

\item{ylim}{a numeric setting the max values of y that will be visualized in the boxplot (zoom only, no data loss).}
}
\value{
a named list containing 'Summary' data.frame and 'Plot' object.
}
\description{
A function that provides an overview of the expression of each marker in the two classes of the dataset.
}
\details{
This function returns a named list that contains two objects:
\itemize{
\item ‘Plot’: a boxplot whose y max value can be set, in order to allow a better visualization (zoom only, no data loss).
\item ‘Summary’: a data.frame with a summary statics of the overall expression of markers in the two classes of the dataset. It  a summary statistics for each class.
}
}
