% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_reports.R
\name{ROC_reports}
\alias{ROC_reports}
\title{Train logistic regression and compute ROC.}
\usage{
ROC_reports(
  data,
  markers_table,
  selected_combinations = NULL,
  single_markers = NULL,
  case_class
)
}
\arguments{
\item{data}{a data.frame returned by load_data().}

\item{markers_table}{a data.frame with ranked combination, reporting: SE, SP, number of markers composing the combination and the score (returned by ranked_combs()).}

\item{selected_combinations}{a numeric vector that specifies the combinations of interest.}

\item{single_markers}{a character vector that specifies the single markers of interest.}

\item{case_class}{a character that specifies which of the two classes of the dataset is the case class.}
}
\value{
a named list containing 3 objects:
\itemize{
\item "Plot": a ggplot object with the ROC curves of the selected combinations.
\item "Metrics": a dataframe with the metrics of the roc curves (AUC, opt. cutoff, etc ...).
\item "Models": the list of models (glm() objects) that have been computed and then used to classify the samples (in which you can find the model equation for each selected combination).
}
}
\description{
A function to compute General Linear Model (binomial) and the corresponding ROC curves for each selected combination.
}
