% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SP.R
\name{SE_SP}
\alias{SE_SP}
\title{Compute sensitivity and specificity of each combination}
\usage{
SE_SP(data, combinations_table)
}
\arguments{
\item{data}{a data.frame returned by load_data().}

\item{combinations_table}{a data.frame containing how many samples of each class are "positive" for each combination (returned by Combi()).}
}
\value{
data.frame with SE, SP and number of composing markers for each combination.
}
\description{
A function to compute sensitivity and specificity of each combination for each class.
}
\details{
This function calculate SE and SP for each combination. The SE of a given combination (capability to find real positives/cases) corresponds to the SE of the case class, while its SP (capability to exclude real negatives/controls) corresponds to the SP of the control class.
}
