% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_markers.R
\name{show_markers}
\alias{show_markers}
\title{Show the composition of combinations.}
\usage{
show_markers(markers_table, selected_combinations)
}
\arguments{
\item{markers_table}{a data.frame with ranked combination, reporting: SE, SP, number of markers composing the combination and the score (returned by ranked_combs()).}

\item{selected_combinations}{a numeric vector that specifies the combinations of interest.}
}
\value{
a data.frame containing the selected combinations and their composing markers.
}
\description{
A function to show the composition of combinations of interest.
}
\examples{
#  A function to show the composition of combinations of interest.

## show_markers(markers_table = ranked_combs_SE_SP, selected_combinations = c(1,11))
}
