% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{signature_score}
\alias{signature_score}
\title{Compute a gene signature score for each cell of a scRNA-seq experiment}
\usage{
signature_score(SeuratObj, geneset, assay = "RNA")
}
\arguments{
\item{SeuratObj}{a SeuratObject}

\item{geneset}{a character containing a list of gene names}

\item{assay}{a character specifying the assay of interest}
}
\value{
a data.frame with 5 columns:
\itemize{
\item coexpression_score: count of how many genes of geneset are expressed in each cell
\item expression_score: represents the overall expression of geneset genes for each cell
\item combined_score: coexpression_score * expression_score
\item scaled_combined_score: combined_score/max(combined_score)
\item log_combined_score: 1/log10(combined_score)
}
}
\description{
Compute a signature score for each cell of a scRNA-seq experiment allowing to vizualize the expression of the whole signature in each cluster.
}
\details{
Given a geneset, this function computes a score for each cell which increases both when the number of expressed (>0) genes in the geneset increases, and when the expression of these genes increases.
This function replicates the code used to compute the "gene signature score" as described in Della Chiara, Gervasoni, Fakiola, Godano et al. 2021.
(https://www.nature.com/articles/s41467-021-22544-y)
}
