% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{Classify}
\alias{Classify}
\title{Classify data.frames using glm(link='binomial') models.}
\usage{
Classify(
  unclassified_data,
  Models,
  Metrics,
  Positive_class = 1,
  Negative_class = 0
)
}
\arguments{
\item{unclassified_data}{a data.frame returned by load_unclassified_data().}

\item{Models}{a list of glm() objects returned by ROC_reports().}

\item{Metrics}{a list of data.frame objects containing ROC metrics, returned by ROC_reports().}

\item{Positive_class}{a numeric or a character that specifies the label of the samples that will be classified as positives}

\item{Negative_class}{a numeric or a character that specifies the label of the samples that will be classified as negatives}
}
\value{
a named list of data.frames, one for each marker/combination contained in the list of models, containg the predicted class for each sample
}
\description{
A function that applies the previously calculated models to an unclassified dataset and classifies the samples.
The unclassified dataset to be classified should be loaded with load_unclassified_data() and MUST contain all the markers of the classified dataset used to train the models (the one loaded with load()).
}
