% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombiROC_long.R
\name{CombiROC_long}
\alias{CombiROC_long}
\title{Reshape CombiROC data in long format.}
\usage{
CombiROC_long(data)
}
\arguments{
\item{data}{a data.frame returned by load_data().}
}
\value{
a data.frame in long format
}
\description{
A function that simply wraps dyplr::pivot_longer() to reshape data in long format.
}
