% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_markers.R
\name{show_markers}
\alias{show_markers}
\title{Show the composition of combinations.}
\usage{
show_markers(markers_table, selected_combinations)
}
\arguments{
\item{markers_table}{a data.frame with combinations returned by Combi().}

\item{selected_combinations}{a numeric vector that specifies the combinations of interest.}
}
\value{
a data.frame containing the selected combinations and their composing markers.
}
\description{
A function to show the composition of combinations of interest.
}
\examples{
demo_data # combiroc built-in demo data (proteomics data from Zingaretti et al. 2012 - PMC3518104)

combs <- combi(data= demo_data, signalthr=450, combithr=1)  # compute combinations

combs_SE_SP <- se_sp(data=demo_data, combinations_table=combs) # compute SE and SP
# of each combination

ranked_combs_SE_SP <- ranked_combs(data= demo_data, combo_table= combs_SE_SP,
                                   case_class='A', min_SE=40, min_SP=80) # rank combinations



#  To show the composition of combinations of interest.

show_markers(markers_table = ranked_combs_SE_SP, selected_combinations = c(1,11))
}
